requirements.txt
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
requests>=2.32
python-dotenv>=1.0